requirements.txt
playwright==1.55.0
pytest==8.3.3
pytest-playwright==0.5.1
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pytest-base-url==2.1.0